omnisia_web/requirements.txt
fastapi
uvicorn
streamlit
requests
ocrmypdf
whisper
Pillow
transformers
sentence-transformers
faiss-cpu
peft
datasets
torch
torchaudio
pydantic
python-multipart
aiofiles
pytesseract
